[BugFix] Use long for ConnectorTableId to avoid int overflow producing negative IDs (#73344)
(cherry picked from commit ac52046e7ce56194f358da222b52cef1d2fee745)

# Conflicts:
#	fe/fe-core/src/main/java/com/starrocks/connector/benchmark/BenchmarkMetadata.java
#	fe/fe-core/src/main/java/com/starrocks/connector/jdbc/JDBCMetadata.java
#	fe/fe-core/src/main/java/com/starrocks/connector/metadata/iceberg/IcebergPropertiesTable.java
#	fe/fe-core/src/main/java/com/starrocks/connector/paimon/PaimonMetadata.java

+++ fe/fe-core/src/main/java/com/starrocks/analysis/TupleDescriptor.java
@@ -169,7 +169,7 @@ public class TupleDescriptor {
         ttupleDesc.setNumNullBytes(-1);
         ttupleDesc.setNumNullSlots(-1);
         if (table != null && table.getId() >= 0) {
-            ttupleDesc.setTableId((int) table.getId());
+            ttupleDesc.setTableId(table.getId());
         }
         return ttupleDesc;
     }
